requirements.txt
streamlit
python-dotenv
mistralai